Fix -fsanitize=undefined diagnostics
Fixes the following diagnostics:
```
bigint.c:631:39: runtime error: left shift of 162 by 24 places cannot be represented in type 'int'
sha1.c:124:42: runtime error: left shift of 152 by 24 places cannot be represented in type 'int'
```

Directs the sanitizer not to instrument the function where this diagnostic
otherwise occurs:
```
crypto_misc.c:131:5: runtime error: load of address 0x7ffd83a8633c with insufficient space for an object of type 'unsigned char'
```
.. this function deliberately reads memory in a way that is undefined
under the C standard, but is a semi-recognized technique to maybe supplement
a CSPRNG entropy pool; blindly removing it is the riskier option, a la
https://www.debian.org/security/2008/dsa-1571

These changes fix all the problems seen in axtls when running the testsuite
of a popular embedded python implementation on a linux amd64 desktop
with gcc 8.3.0 -fsanitize=undefined.

+++ crypto/bigint.c
@@ -628,7 +628,7 @@ bigint *bi_import(BI_CTX *ctx, const uint8_t *data, int size)
 
     for (i = size-1; i >= 0; i--)
     {
-        biR->comps[offset] += data[i] << (j*8);
+        biR->comps[offset] += (unsigned int)data[i] << (j*8);
 
         if (++j == COMP_BYTE_SIZE)
         {
